Small refactors to test.
