+++ .nvmrc
@@ -1,1 +1,1 @@
-v18.8.0
+16.20.2
