+++ pages/index.tsx
@@ -1,6 +1,5 @@
-import type { InferGetStaticPropsType, GetStaticProps } from 'next'
-import Head from 'next/head'
 import dynamic from 'next/dynamic'
+import AppHead from '../components/AppHead'
 import Navbar from '../components/Navbar'
 import MainHero from '../components/MainHero'
 import Footer from '../components/Footer'
@@ -23,37 +22,12 @@ const Testimonials = dynamic(() => import("../components/Testimonials"), {
    ssr: false,
 });
 
-export function AppHead() {
-  return (
-    <Head>
-      <title>
-        Sustainable Progress Equality Collective
-      </title>
-      <link rel="icon" type="image/png" href="/favicon-16x16.png" />
-      <meta property="og:title" content="Sustainable Progress and Equality Collective" />
-      <meta property="og:url" content="https://specollective.org/" />
-      <meta
-        property="og:image"
-        content="https://spec-bucket.nyc3.cdn.digitaloceanspaces.com/spec-logo-old.png"
-      />
-      <meta
-        property="og:description"
-        content="Helping people learn skills, build careers, and become leaders of sustainable social impact."
-      />
-      <meta
-        name="twitter:card"
-        content="https://spec-bucket.nyc3.cdn.digitaloceanspaces.com/spec-logo-old.png"
-      />
-      <meta name="twitter:title" content="Sustainable Progress and Equality Collective" />
-    </Head>
-  )
-}
-
 type HomeProps = {
   title: string;
   heroContent: string;
 };
 
+// Langing page for app.
 export default function Home({ title, heroContent }: HomeProps) {
   return (
     <>
